refactor: TargetField 분리
